requirements.txt
nameko~=2.13.0
SQLAlchemy~=1.4.22
alembic~=1.6.5
python-dotenv~=0.19.0
psycopg2-binary~=2.9.1
